requirements.txt
matplotlib.pyplot
seaborn 
streamlit 
pandas
seaborn 
emoji
collections 
wordcloud 
urlextract 
re